# frontend/types/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Swell Travel Agent API",
  "description": "Surf-forecast-driven travel discovery. All state-changing requests need the session cookie and an `X-CSRF-Token` header matching the `sta_csrf` cookie (GET `/api/auth/csrf` first).",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/searches": {
   "get": {
    "tags": [
     "searches"
    ],
    "summary": "List Searches",
    "operationId": "list_searches_api_searches_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SearchOut"
         },
         "type": "array",
         "title": "Response List Searches Api Searches Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "searches"
    ],
    "summary": "Create",
    "operationId": "create_api_searches_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SearchIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/searches/{search_id}": {
   "get": {
    "tags": [
     "searches"
    ],
    "summary": "Get",
    "operationId": "get_api_searches__search_id__get",
    "parameters": [
     {
      "name": "search_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Search Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "searches"
    ],
    "summary": "Update",
    "operationId": "update_api_searches__search_id__put",
    "parameters": [
     {
      "name": "search_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Search Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SearchIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "searches"
    ],
    "summary": "Delete",
    "operationId": "delete_api_searches__search_id__delete",
    "parameters": [
     {
      "name": "search_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Search Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Message"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/searches/{search_id}/pause": {
   "post": {
    "tags": [
     "searches"
    ],
    "summary": "Pause",
    "operationId": "pause_api_searches__search_id__pause_post",
    "parameters": [
     {
      "name": "search_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Search Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/searches/{search_id}/resume": {
   "post": {
    "tags": [
     "searches"
    ],
    "summary": "Resume",
    "operationId": "resume_api_searches__search_id__resume_post",
    "parameters": [
     {
      "name": "search_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Search Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/searches/{search_id}/duplicate": {
   "post": {
    "tags": [
     "searches"
    ],
    "summary": "Duplicate",
    "operationId": "duplicate_api_searches__search_id__duplicate_post",
    "parameters": [
     {
      "name": "search_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Search Id"
      }
     }
    ],
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/searches/{search_id}/matches": {
   "get": {
    "tags": [
     "searches"
    ],
    "summary": "Matches",
    "operationId": "matches_api_searches__search_id__matches_get",
    "parameters": [
     {
      "name": "search_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Search Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MatchSummary"
         },
         "title": "Response Matches Api Searches  Search Id  Matches Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BreakType": {
    "type": "string",
    "enum": [
     "reef",
     "point",
     "beach",
     "rivermouth",
     "slab"
    ],
    "title": "BreakType"
   },
   "CabinClass": {
    "type": "string",
    "enum": [
     "economy",
     "premium_economy",
     "business",
     "first"
    ],
    "title": "CabinClass"
   },
...
   "DateMode": {
    "type": "string",
    "enum": [
     "fixed",
     "flexible"
    ],
    "title": "DateMode"
   },
...
   "DestinationMode": {
    "type": "string",
    "enum": [
     "all",
     "regions",
     "spots"
    ],
    "title": "DestinationMode"
   },
   "DestinationsIn": {
    "properties": {
     "spot_slugs": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 60,
      "title": "Spot Slugs"
     },
     "countries": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 60,
      "title": "Countries"
     },
     "region_groups": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 20,
      "title": "Region Groups"
     }
    },
    "type": "object",
    "title": "DestinationsIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MatchCounts": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total",
      "default": 0
     },
     "flight_found": {
      "type": "integer",
      "title": "Flight Found",
      "default": 0
     },
     "surf_only": {
      "type": "integer",
      "title": "Surf Only",
      "default": 0
     },
     "pending_flights": {
      "type": "integer",
      "title": "Pending Flights",
      "default": 0
     },
     "expired": {
      "type": "integer",
      "title": "Expired",
      "default": 0
     }
    },
    "type": "object",
    "title": "MatchCounts"
   },
...
   "MatchSummary": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "search_id": {
      "type": "string",
      "format": "uuid",
      "title": "Search Id"
     },
     "search_name": {
      "type": "string",
      "title": "Search Name"
     },
     "search_status": {
      "type": "string",
      "title": "Search Status"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "status_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status Reason"
     },
     "spot": {
      "$ref": "#/components/schemas/SpotMini"
     },
     "event_id": {
      "type": "string",
      "format": "uuid",
      "title": "Event Id"
     },
     "event_status": {
      "type": "string",
      "title": "Event Status"
     },
     "window_start": {
      "type": "string",
      "format": "date-time",
      "title": "Window Start"
     },
     "window_end": {
      "type": "string",
      "format": "date-time",
      "title": "Window End"
     },
     "qualifying_hours": {
      "type": "integer",
      "title": "Qualifying Hours"
     },
     "peak_score": {
      "type": "integer",
      "title": "Peak Score"
     },
     "avg_score": {
      "type": "number",
      "title": "Avg Score"
     },
     "peak_label": {
      "type": "string",
      "title": "Peak Label"
     },
     "breaking_height_min_ft": {
      "type": "number",
      "title": "Breaking Height Min Ft"
     },
     "breaking_height_max_ft": {
      "type": "number",
      "title": "Breaking Height Max Ft"
     },
     "confidence": {
      "type": "integer",
      "title": "Confidence"
     },
     "confidence_label": {
      "type": "string",
      "title": "Confidence Label"
     },
     "destination_iata": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Destination Iata"
     },
     "transfer_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Transfer Minutes"
     },
     "origin_iata": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Origin Iata"
     },
     "recommended_arrival_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommended Arrival Date"
     },
     "recommended_departure_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommended Departure Date"
     },
     "best_price": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Best Price"
     },
     "currency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency"
     },
     "overall_score": {
      "type": "integer",
      "title": "Overall Score"
     },
     "scores": {
      "additionalProperties": true,
      "type": "object",
      "title": "Scores"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "has_mock_flights": {
      "type": "boolean",
      "title": "Has Mock Flights"
     },
     "flight_checked_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Flight Checked At"
     },
     "notification_count": {
      "type": "integer",
      "title": "Notification Count"
     },
     "lead_days": {
      "type": "number",
      "title": "Lead Days"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "search_id",
     "search_name",
     "search_status",
     "status",
     "status_reason",
     "spot",
     "event_id",
     "event_status",
     "window_start",
     "window_end",
     "qualifying_hours",
     "peak_score",
     "avg_score",
     "peak_label",
     "breaking_height_min_ft",
     "breaking_height_max_ft",
     "confidence",
     "confidence_label",
     "destination_iata",
     "transfer_minutes",
     "origin_iata",
     "recommended_arrival_date",
     "recommended_departure_date",
     "best_price",
     "currency",
     "overall_score",
     "scores",
     "is_demo",
     "has_mock_flights",
     "flight_checked_at",
     "notification_count",
     "lead_days",
     "created_at",
     "updated_at"
    ],
    "title": "MatchSummary"
   },
   "Message": {
    "properties": {
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "message"
    ],
    "title": "Message"
   },
   "NotificationChannelPref": {
    "type": "string",
    "enum": [
     "email",
     "sms",
     "both",
     "dashboard"
    ],
    "title": "NotificationChannelPref"
   },
...
   "QualityLabel": {
    "type": "string",
    "enum": [
     "poor",
     "fair",
     "good",
     "very_good",
     "excellent",
     "exceptional"
    ],
    "title": "QualityLabel"
   },
   "RankingPriority": {
    "type": "string",
    "enum": [
     "balanced",
     "best_waves",
     "cheapest",
     "shortest"
    ],
    "title": "RankingPriority"
   },
...
   "SearchIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Name"
     },
     "date_mode": {
      "$ref": "#/components/schemas/DateMode",
      "default": "flexible"
     },
     "date_start": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date Start"
     },
     "date_end": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date End"
     },
     "horizon_days": {
      "type": "integer",
      "maximum": 365.0,
      "minimum": 1.0,
      "title": "Horizon Days",
      "default": 30
     },
     "units": {
      "type": "string",
      "pattern": "^(ft|m)$",
      "title": "Units",
      "default": "ft"
     },
     "wave_min": {
      "type": "number",
      "maximum": 100.0,
      "minimum": 0.0,
      "title": "Wave Min"
     },
     "wave_max": {
      "type": "number",
      "maximum": 100.0,
      "exclusiveMinimum": 0.0,
      "title": "Wave Max"
     },
     "wave_preferred": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Wave Preferred"
     },
     "min_quality": {
      "$ref": "#/components/schemas/QualityLabel",
      "default": "good"
     },
     "min_period_s": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 25.0,
        "minimum": 4.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Period S"
     },
     "max_wind_kmh": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Wind Kmh"
     },
     "wind_requirement": {
      "$ref": "#/components/schemas/WindRequirement",
      "default": "any"
     },
     "swell_direction_min": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMaximum": 360.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Swell Direction Min"
     },
     "swell_direction_max": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMaximum": 360.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Swell Direction Max"
     },
     "min_consistency": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Consistency"
     },
     "break_types": {
      "items": {
       "$ref": "#/components/schemas/BreakType"
      },
      "type": "array",
      "title": "Break Types"
     },
     "min_window_hours": {
      "type": "integer",
      "maximum": 240.0,
      "minimum": 1.0,
      "title": "Min Window Hours",
      "default": 4
     },
     "destination_mode": {
      "$ref": "#/components/schemas/DestinationMode",
      "default": "all"
     },
     "destinations": {
      "$ref": "#/components/schemas/DestinationsIn"
     },
     "max_transfer_minutes": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 2880.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Transfer Minutes"
     },
     "origins": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 5,
      "minItems": 1,
      "title": "Origins"
     },
     "max_origin_ground_km": {
      "type": "integer",
      "maximum": 500.0,
      "minimum": 0.0,
      "title": "Max Origin Ground Km",
      "default": 0
     },
     "notification_channel": {
      "$ref": "#/components/schemas/NotificationChannelPref",
      "default": "email"
     },
     "notify_surf_only": {
      "type": "boolean",
      "title": "Notify Surf Only",
      "default": false
     },
     "notify_on_updates": {
      "type": "boolean",
      "title": "Notify On Updates",
      "default": true
     },
     "priority": {
      "$ref": "#/components/schemas/RankingPriority",
      "default": "balanced"
     },
     "travel": {
      "$ref": "#/components/schemas/TravelIn"
     }
    },
    "type": "object",
    "required": [
     "name",
     "wave_min",
     "wave_max",
     "origins",
     "travel"
    ],
    "title": "SearchIn"
   },
   "SearchOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "status": {
      "$ref": "#/components/schemas/SearchStatus"
     },
     "date_mode": {
      "$ref": "#/components/schemas/DateMode"
     },
     "date_start": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date Start"
     },
     "date_end": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date End"
     },
     "horizon_days": {
      "type": "integer",
      "title": "Horizon Days"
     },
     "units": {
      "type": "string",
      "title": "Units"
     },
     "wave_min": {
      "type": "number",
      "title": "Wave Min"
     },
     "wave_max": {
      "type": "number",
      "title": "Wave Max"
     },
     "wave_preferred": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Wave Preferred"
     },
     "wave_min_ft": {
      "type": "number",
      "title": "Wave Min Ft"
     },
     "wave_max_ft": {
      "type": "number",
      "title": "Wave Max Ft"
     },
     "min_quality": {
      "$ref": "#/components/schemas/QualityLabel"
     },
     "min_period_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Period S"
     },
     "max_wind_kmh": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Wind Kmh"
     },
     "wind_requirement": {
      "$ref": "#/components/schemas/WindRequirement"
     },
     "swell_direction_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Swell Direction Min"
     },
     "swell_direction_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Swell Direction Max"
     },
     "min_consistency": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Consistency"
     },
     "break_types": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Break Types"
     },
     "min_window_hours": {
      "type": "integer",
      "title": "Min Window Hours"
     },
     "destination_mode": {
      "$ref": "#/components/schemas/DestinationMode"
     },
     "destinations": {
      "$ref": "#/components/schemas/DestinationsIn"
     },
     "max_transfer_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Transfer Minutes"
     },
     "origins": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Origins"
     },
     "max_origin_ground_km": {
      "type": "integer",
      "title": "Max Origin Ground Km"
     },
     "notification_channel": {
      "$ref": "#/components/schemas/NotificationChannelPref"
     },
     "notify_surf_only": {
      "type": "boolean",
      "title": "Notify Surf Only"
     },
     "notify_on_updates": {
      "type": "boolean",
      "title": "Notify On Updates"
     },
     "priority": {
      "$ref": "#/components/schemas/RankingPriority"
     },
     "travel": {
      "$ref": "#/components/schemas/TravelOut"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "paused_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Paused At"
     },
     "last_evaluated_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Evaluated At"
     },
     "match_counts": {
      "$ref": "#/components/schemas/MatchCounts"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "status",
     "date_mode",
     "date_start",
     "date_end",
     "horizon_days",
     "units",
     "wave_min",
     "wave_max",
     "wave_preferred",
     "wave_min_ft",
     "wave_max_ft",
     "min_quality",
     "min_period_s",
     "max_wind_kmh",
     "wind_requirement",
     "swell_direction_min",
     "swell_direction_max",
     "min_consistency",
     "break_types",
     "min_window_hours",
     "destination_mode",
     "destinations",
     "max_transfer_minutes",
     "origins",
     "max_origin_ground_km",
     "notification_channel",
     "notify_surf_only",
     "notify_on_updates",
     "priority",
     "travel",
     "created_at",
     "updated_at",
     "paused_at",
     "last_evaluated_at",
     "match_counts"
    ],
    "title": "SearchOut"
   },
   "SearchStatus": {
    "type": "string",
    "enum": [
     "active",
     "paused"
    ],
    "title": "SearchStatus"
   },
...
   "SpotMini": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "country": {
      "type": "string",
      "title": "Country"
     },
     "country_code": {
      "type": "string",
      "title": "Country Code"
     },
     "region": {
      "type": "string",
      "title": "Region"
     },
     "region_group": {
      "type": "string",
      "title": "Region Group"
     },
     "latitude": {
      "type": "number",
      "title": "Latitude"
     },
     "longitude": {
      "type": "number",
      "title": "Longitude"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone"
     },
     "break_type": {
      "type": "string",
      "title": "Break Type"
     }
    },
    "type": "object",
    "required": [
     "id",
     "slug",
     "name",
     "country",
     "country_code",
     "region",
     "region_group",
     "latitude",
     "longitude",
     "timezone",
     "break_type"
    ],
    "title": "SpotMini"
   },
...
   "TravelIn": {
    "properties": {
     "max_price": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100000.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "string"
       }
      ],
      "title": "Max Price"
     },
     "currency": {
      "type": "string",
      "title": "Currency",
      "default": "USD"
     },
     "direct_only": {
      "type": "boolean",
      "title": "Direct Only",
      "default": false
     },
     "max_layovers": {
      "type": "integer",
      "maximum": 3.0,
      "minimum": 0.0,
      "title": "Max Layovers",
      "default": 2
     },
     "max_flight_hours": {
      "type": "number",
      "maximum": 60.0,
      "minimum": 1.0,
      "title": "Max Flight Hours",
      "default": 30
     },
     "min_days_at_destination": {
      "type": "integer",
      "maximum": 60.0,
      "minimum": 1.0,
      "title": "Min Days At Destination",
      "default": 3
     },
     "max_trip_days": {
      "type": "integer",
      "maximum": 90.0,
      "minimum": 1.0,
      "title": "Max Trip Days",
      "default": 14
     },
     "travelers": {
      "type": "integer",
      "maximum": 9.0,
      "minimum": 1.0,
      "title": "Travelers",
      "default": 1
     },
     "preferred_airlines": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 10,
      "title": "Preferred Airlines"
     },
     "cabin_class": {
      "$ref": "#/components/schemas/CabinClass",
      "default": "economy"
     },
     "arrival_buffer_days": {
      "type": "integer",
      "maximum": 14.0,
      "minimum": 0.0,
      "title": "Arrival Buffer Days",
      "default": 2
     },
     "departure_buffer_days": {
      "type": "integer",
      "maximum": 14.0,
      "minimum": 0.0,
      "title": "Departure Buffer Days",
      "default": 1
     }
    },
    "type": "object",
    "required": [
     "max_price"
    ],
    "title": "TravelIn"
   },
   "TravelOut": {
    "properties": {
     "max_price": {
      "type": "string",
      "title": "Max Price"
     },
     "currency": {
      "type": "string",
      "title": "Currency",
      "default": "USD"
     },
     "direct_only": {
      "type": "boolean",
      "title": "Direct Only",
      "default": false
     },
     "max_layovers": {
      "type": "integer",
      "maximum": 3.0,
      "minimum": 0.0,
      "title": "Max Layovers",
      "default": 2
     },
     "max_flight_hours": {
      "type": "number",
      "maximum": 60.0,
      "minimum": 1.0,
      "title": "Max Flight Hours",
      "default": 30
     },
     "min_days_at_destination": {
      "type": "integer",
      "maximum": 60.0,
      "minimum": 1.0,
      "title": "Min Days At Destination",
      "default": 3
     },
     "max_trip_days": {
      "type": "integer",
      "maximum": 90.0,
      "minimum": 1.0,
      "title": "Max Trip Days",
      "default": 14
     },
     "travelers": {
      "type": "integer",
      "maximum": 9.0,
      "minimum": 1.0,
      "title": "Travelers",
      "default": 1
     },
     "preferred_airlines": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 10,
      "title": "Preferred Airlines"
     },
     "cabin_class": {
      "$ref": "#/components/schemas/CabinClass",
      "default": "economy"
     },
     "arrival_buffer_days": {
      "type": "integer",
      "maximum": 14.0,
      "minimum": 0.0,
      "title": "Arrival Buffer Days",
      "default": 2
     },
     "departure_buffer_days": {
      "type": "integer",
      "maximum": 14.0,
      "minimum": 0.0,
      "title": "Departure Buffer Days",
      "default": 1
     }
    },
    "type": "object",
    "required": [
     "max_price"
    ],
    "title": "TravelOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WindRequirement": {
    "type": "string",
    "enum": [
     "any",
     "not_onshore",
     "offshore"
    ],
    "title": "WindRequirement"
   }
  }
 }
}